+++ src/main/java/wooteco/subway/service/dto/LineRequest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.dto;
+package wooteco.subway.service.dto;
 
 import wooteco.subway.domain.Line;
