+++ src/main/java/wooteco/subway/ui/StationController.java
@@ -2,8 +2,8 @@ package wooteco.subway.ui;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import wooteco.subway.dto.StationRequest;
-import wooteco.subway.dto.StationResponse;
+import wooteco.subway.service.dto.StationRequest;
+import wooteco.subway.service.dto.StationResponse;
 import wooteco.subway.service.StationService;
 
 import java.net.URI;
